refactor(frontend): extract helpers batch 14 (#244)
Extract data arrays, types, and interfaces into co-located helper files.
- showcase.tsx: extract VerticalPreview + verticals data → showcase-helpers.ts
- features-grid.tsx: extract FEATURES array + Feature type → features-grid-helpers.ts
- TenantContext.tsx: extract 10 interfaces + PUBLIC_PATHS → tenant-context-types.ts
