+++ frontend/src/components/marketing/features-grid.tsx
@@ -4,58 +4,9 @@ import { useRef } from 'react';
 import { gsap } from 'gsap';
 import { ScrollTrigger } from 'gsap/ScrollTrigger';
 import { useGSAP } from '@gsap/react';
-import {
-  ShoppingBag,
-  CalendarCheck,
-  BarChart3,
-  Star,
-  TrendingUp,
-} from 'lucide-react';
+import { FEATURES } from './features-grid-helpers';
 
 gsap.registerPlugin(ScrollTrigger);
-
-const FEATURES = [
-  {
-    icon: ShoppingBag,
-    title: 'Tienda online',
-    description: 'Catálogo, carrito y pagos integrados. Vende desde el primer día.',
-    visual: 'store',
-    iconBg: 'bg-primary/10',
-    iconText: 'text-primary',
-  },
-  {
-    icon: CalendarCheck,
-    title: 'Reservas',
-    description: 'Tus clientes reservan solos. Sin llamadas, sin WhatsApp.',
-    visual: 'calendar',
-    iconBg: 'bg-emerald-500/10',
-    iconText: 'text-emerald-600 dark:text-emerald-400',
-  },
-  {
-    icon: TrendingUp,
-    title: 'Marketing',
-    description: 'Campañas y notificaciones automáticas que traen clientes de vuelta.',
-    visual: 'chart-bars',
-    iconBg: 'bg-violet-500/10',
-    iconText: 'text-violet-600 dark:text-violet-400',
-  },
-  {
-    icon: Star,
-    title: 'Reseñas',
-    description: 'Reseñas reales que generan confianza y mejoran tu posición en Google.',
-    visual: 'reviews',
-    iconBg: 'bg-amber-500/10',
-    iconText: 'text-amber-600 dark:text-amber-400',
-  },
-  {
-    icon: BarChart3,
-    title: 'Analíticas',
-    description: 'Métricas claras de visitas, ventas y rendimiento. Sin complicaciones.',
-    visual: 'chart-line',
-    iconBg: 'bg-sky-500/10',
-    iconText: 'text-sky-600 dark:text-sky-400',
-  },
-];
 
 // ─── Mini visuals for each feature card ─────────────────
